Split Qwen35 virtual KV reporting
